+++ src/Config/route-architect.php
@@ -1,0 +1,15 @@
+<?php
+
+return [
+	
+	/*
+	|--------------------------------------------------------------------------
+	| Auto Scan
+	|--------------------------------------------------------------------------
+	|
+	| The option determines whether routes should be declaring automatically.
+	|
+	*/
+
+    'auto_scan' => env('ROUTE_ARCHITECT_AUTO_SCAN', false),
+];