+++ src/Providers/RouteArchitectServiceProvider.php
@@ -1,6 +1,7 @@
 <?php
 
-namespace TeaAroma\RouteArchitect;
+namespace TeaAroma\RouteArchitect\Providers;
+
 
 use Illuminate\Support\ServiceProvider;
 
@@ -15,5 +16,8 @@ class RouteArchitectServiceProvider extends ServiceProvider
 	/**
 	 * @return void
 	 */
-	public function boot(): void {}
+	public function boot(): void
+	{
+		$this->publishes([ __DIR__ . '/../Config/route-architect.php' => config_path('route-architect.php') ], 'config');
+	}
 }