+++ CHANGELOG.md
@@ -9,3 +9,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Added
 
 - Initial creation of the project.
+
+## [0.0.2] - 2025-04-15
+
+### Added
+
+- Abstract class `RouteArchitect` providing base core routing logic and structure.
+- `IsNotMiddleware` callable for filtering specific middlewares.
+- Base configuration file for customizable behavior of this package.
+- `RouteArchitectTypes` enum encapsulating supported method types of `type` property of the `RouteArchitect` class.
+- `RouteArchitectHelpers` with base utility functions.